+++ client/pages/forgot-password.vue
@@ -130,11 +130,17 @@ export default {
       return self.$nextTick(() => self.$router.go(-1))
     },
     async sendEmail () {
-      const { data } = await this.form.post('/password/email')
+      if (this.$refs.form.validate()) {
+        this.form.busy = true
 
-      this.status = data.status
+        const { data } = await this.form.post('/password/email')
 
-      this.form.reset()
+        this.status = data.status
+
+        this.form.reset()
+
+        this.form.busy = false
+      }
     }
   }